[Database] Huge update for all modes
This commit introduces structure GameInfo that collects information about the
game before, during, and after it. It is then used to save much many values
related to game results into database.

Results table structure was done from scratch, and if you try to run with an
older version of the database, it will tell you either to update it, or to
disable the warning. Old CMake parameters USE_RECORDS_V* and ENABLE_RECORDS_V*
are removed. Databases are expected to adhere to the new structure, which is
not supposed to be changed for a long time.

Docs folder contains the description of the database changes, and how to update
the tables to the latest version.

To obtain more info about the game, some classes have to be changed a bit. In
particular, soccer goals and CTF captures are now stored in the database, and
the info about players joining and leaving the game.

What should be done later in addition to these changes?
* Adding the ability to configure what should be stored in the database -
not everyone wants soccer/ctf information, and for some settings like
time-limited soccer that could be too many data saved.
* Logging the failed SQL queries (for retrying to invoke them)
* More documentation
* Some code refactoring will happen