Move the teamhit logic away from server lobby
I would like items to not refer to server lobby too, but the structure of classes interaction was written by very nice people who didn't think I would ever need to pass one pointer from ServerLobby to Bowling, and so it's pretty difficult and it's not happening now.
Very sad, because I just want to not include SL in many other files.